+++ view/tws_administration_view/lib/view/articles/drivers/whispers/forms/drivers/drivers_create_address_form.dart
@@ -70,7 +70,7 @@ class _DriversCreateAddressForm extends StatelessWidget {
                       ) ??
                       Employee.a().clone(
                         addressNavigation: Address.a().clone(
-                          state: text,
+                          country: text,
                         ),
                       ),
                     ),